+++ src/engine/sparqlExpressions/SparqlExpression.cpp
@@ -61,7 +61,7 @@ auto SparqlExpression::isAggregate() const -> AggregateStatus {
 // _____________________________________________________________________________
 std::unique_ptr<SparqlExpression> SparqlExpression::replaceChild(
     size_t childIndex, std::unique_ptr<SparqlExpression> newExpression) {
-  AD_CONTRACT_CHECK(childIndex < static_cast<size_t>(children().size()));
+  AD_CONTRACT_CHECK(childIndex < children().size());
   return std::exchange(children()[childIndex], std::move(newExpression));
 }
