+++ src/engine/sparqlExpressions/SparqlExpressionGenerators.h
@@ -31,7 +31,7 @@ inline ql::span<const ValueId> getIdsFromVariable(
   ql::span<const ValueId> completeColumn = inputTable.getColumn(columnIndex);
 
   AD_CONTRACT_CHECK(beginIndex <= endIndex &&
-                    endIndex <= static_cast<size_t>(completeColumn.size()));
+                    endIndex <= completeColumn.size());
   return {completeColumn.begin() + beginIndex,
           completeColumn.begin() + endIndex};
 }
@@ -65,7 +65,7 @@ CPP_template(typename T, typename Transformation = std::identity)(
     requires ql::ranges::input_range<
         T>) auto resultGenerator(T&& vector, size_t numItems,
                                  Transformation transformation = {}) {
-  AD_CONTRACT_CHECK(numItems == static_cast<size_t>(vector.size()));
+  AD_CONTRACT_CHECK(numItems == vector.size());
   return ad_utility::allView(AD_FWD(vector)) |
          ql::views::transform(std::move(transformation));
 }
